# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Inventory Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/sales-summary:
    get:
      operationId: getSalesSummaryReport
      tags: [reports]
      summary: Sales summary report
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, weekly, monthly, annual]
            default: monthly
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Sales summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesSummaryReport"
  /reports/purchase-summary:
    get:
      operationId: getPurchaseSummaryReport
      tags: [reports]
      summary: Purchase summary report
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Purchase summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseSummaryReport"
  /reports/inventory:
    get:
      operationId: getInventoryReport
      tags: [reports]
      summary: Inventory report
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [current, low-stock, out-of-stock, dead-stock, valuation]
            default: current
        - name: categoryId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Inventory report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryReport"
  /reports/gst-summary:
    get:
      operationId: getGstSummaryReport
      tags: [reports]
      summary: GST summary report
      security:
        - bearerAuth: []
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: integer
        - name: year
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: GST summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GstSummaryReport"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SalesSummaryReport:
      type: object
      properties:
        totalSales:
          type: number
        totalTax:
          type: number
        totalDiscount:
          type: number
        netAmount:
          type: number
        transactionCount:
          type: integer
        breakdown:
          type: array
          items:
            $ref: "#/components/schemas/SalesSummaryBreakdown"
      required:
        - totalSales
        - totalTax
        - totalDiscount
        - netAmount
        - transactionCount
        - breakdown
    SalesSummaryBreakdown:
      type: object
      properties:
        period:
          type: string
        sales:
          type: number
        tax:
          type: number
        count:
          type: integer
      required:
        - period
        - sales
        - tax
        - count
    PurchaseSummaryReport:
      type: object
      properties:
        totalPurchases:
          type: number
        totalTax:
          type: number
        totalDiscount:
          type: number
        netAmount:
          type: number
        transactionCount:
          type: integer
        pendingPayments:
          type: number
      required:
        - totalPurchases
        - totalTax
        - totalDiscount
        - netAmount
        - transactionCount
        - pendingPayments
    InventoryReport:
      type: object
      properties:
        totalItems:
          type: integer
        totalValue:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/InventoryReportItem"
      required:
        - totalItems
        - totalValue
        - items
    InventoryReportItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sku:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        currentStock:
          type: number
        unit:
          type: string
        purchasePrice:
          type: number
        sellingPrice:
          type: number
        stockValue:
          type: number
        minStockAlert:
          type: number
        lastMovement:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - name
        - currentStock
        - unit
        - purchasePrice
        - sellingPrice
        - stockValue
        - minStockAlert
    GstSummaryReport:
      type: object
      properties:
        month:
          type: integer
        year:
          type: integer
        outputTax:
          type: number
        inputTax:
          type: number
        netPayable:
          type: number
        salesBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/GstRateBreakdown"
        purchaseBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/GstRateBreakdown"
      required:
        - month
        - year
        - outputTax
        - inputTax
        - netPayable
        - salesBreakdown
        - purchaseBreakdown
    GstRateBreakdown:
      type: object
      properties:
        gstRate:
          type: number
        taxableAmount:
          type: number
        cgst:
          type: number
        sgst:
          type: number
        igst:
          type: number
        totalTax:
          type: number
      required:
        - gstRate
        - taxableAmount
        - cgst
        - sgst
        - igst
        - totalTax
...